changing the layout of the tests into a new package name to match the dir and breakout utility scripts
