+++ tests/chmgt_test.go
@@ -1,27 +1,17 @@
-package chmgt_test
+package tests
 
 import (
-	"bytes"
-	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"net/http"
-	"net/http/httptest"
 	"os"
 	"testing"
 
 	"github.com/heimdal-rw/chmgt/config"
 	"github.com/heimdal-rw/chmgt/handling"
 	"github.com/heimdal-rw/chmgt/models"
+	"github.com/stretchr/testify/assert"
 )
-
-type App struct {
-	Router http.Handler
-	DB     *models.Datasource
-}
-
-var a App
 
 func TestMain(m *testing.M) {
 	a = App{}
@@ -34,10 +24,7 @@ func TestMain(m *testing.M) {
 	if err != nil {
 		log.Fatal(err)
 	}
-	handler.Datasource, err = models.NewDatasource(
-		conf.DatabaseConnection(),
-		conf.Database.Name,
-	)
+	handler.Datasource, err = models.NewDatasource(conf)
 
 	a.Router = handler.Router
 	a.DB = handler.Datasource
@@ -58,9 +45,7 @@ func TestGetSingleUser(t *testing.T) {
 		t.Errorf("Got %s when attempting to auth", err)
 	}
 
-	req, _ := http.NewRequest("GET", "/api/users", nil)
-	req.Header.Add("Authorization", fmt.Sprintf("Bearer %s", token))
-	response := executeRequest(req)
+	response := executeRequest("GET", "/api/users", token, nil)
 
 	checkResponseCode(t, http.StatusOK, response.Code)
 
@@ -69,46 +54,42 @@ func TestGetSingleUser(t *testing.T) {
 		t.Errorf("Error in response format. Got %s", err)
 	}
 
-	if fmtresp.Success != true {
-		t.Errorf("Expected success to be 'true'. Got %v", fmtresp.Success)
-	}
-
-	if fmtresp.Message != "success" {
-		t.Errorf("Expected message to be 'success'. Got %v", fmtresp.Message)
-	}
-
+	// Testing the response object for this particular function
 	body := formatData(fmtresp.Data)
-
-	if len(body) > 1 {
-		t.Errorf("Expected the body to only have 1 hit. Got %v", len(body))
-	}
-
-	if val, ok := body[0]["password"]; ok {
-		t.Errorf("Expected no password to be returned. Got %v", val)
-	}
-
-	if body[0]["username"] != "admin" {
-		t.Errorf("Expected the username to be 'admin'. Got %v", body[0]["username"])
-	}
-
-	if body[0]["email"] != "admin@example.com" {
-		t.Errorf("Expected the username to be 'admin@example.com'. Got %v", body[0]["email"])
-	}
-
-	if body[0]["firstname"] != "admin" {
-		t.Errorf("Expected the firstname to be 'Admin'. Got %v", body[0]["firstname"])
-	}
-
-	if body[0]["lastname"] != "User" {
-		t.Errorf("Expected the lastname to be 'User'. Got %v", body[0]["firstname"])
-	}
+	assert.Exactly(t, bool(true), fmtresp.Success, fmt.Sprintf("Expected 'success' object to be true. Got: %v", fmtresp.Success))
+	assert.Exactly(t, string("success"), fmtresp.Message, fmt.Sprintf("Expected response message to be 'success'. Got: %v", fmtresp.Message))
+	assert.Exactly(t, int(1), len(body), fmt.Sprintf("Expected the body to only have 1 hit. Got: %v", len(body)))
+	assert.Exactly(t, nil, body[0]["password"], fmt.Sprintf("Expected no password to be returned. Got %v", body[0]["password"]))
+	assert.Exactly(t, string("admin"), body[0]["username"], fmt.Sprintf("Expected the username to be 'admin'. Got %v", body[0]["username"]))
+	assert.Exactly(t, string("admin@example.com"), body[0]["email"], fmt.Sprintf("Expected the username to be 'admin@example.com'. Got %v", body[0]["email"]))
+	assert.Exactly(t, string("admin"), body[0]["firstname"], fmt.Sprintf("Expected the firstname to be 'Admin'. Got %v", body[0]["firstname"]))
+	assert.Exactly(t, string("User"), body[0]["lastname"], fmt.Sprintf("Expected the lastname to be 'User'. Got %v", body[0]["lastname"]))
 }
 
 // TODO: Populate tests below
 
-// func TestGetMultipleUsers(t *testing.T) {
+func TestGetMultipleUsers(t *testing.T) {
+	clearCollections()
+	insertUsers(a.DB, []string{"admin", "tester1", "tester2"})
 
-// }
+	token, err := getAuthToken("admin", "password_admin")
+	if err != nil {
+		t.Errorf("Got %s when attempting to auth", err)
+	}
+
+	response := executeRequest("GET", "/api/users", token, nil)
+
+	checkResponseCode(t, http.StatusOK, response.Code)
+
+	fmtresp, err := formatResponse(response)
+	if err != nil {
+		t.Errorf("Error in response format. Got %s", err)
+	}
+
+	// Testing the response object for this particular function
+	body := formatData(fmtresp.Data)
+	assert.Exactly(t, int(3), len(body), fmt.Sprintf("Expected the body to only 3 hits. Got: %v", len(body)))
+}
 
 // func TestGetNonExistingUser(t *testing.T) {
 
@@ -149,75 +130,3 @@ func TestGetSingleUser(t *testing.T) {
 // func TestDeleteChangeRequest(t *testing.T) {
 
 // }
-
-func executeRequest(req *http.Request) *httptest.ResponseRecorder {
-	rr := httptest.NewRecorder()
-	a.Router.ServeHTTP(rr, req)
-
-	return rr
-}
-
-func checkResponseCode(t *testing.T, expected, actual int) {
-	if expected != actual {
-		t.Errorf("Expected response code %d. Got %d\n", expected, actual)
-	}
-}
-
-func getAuthToken(user string, pw string) (string, error) {
-
-	authString := fmt.Sprintf(`{"username": "%s", "password": "%s"}`, user, pw)
-	data := []byte(authString)
-
-	req, err := http.NewRequest("POST", "/api/authenticate", bytes.NewBuffer(data))
-	if err != nil {
-		return "", err
-	}
-	req.Header.Set("Content-Type", "application/json")
-	response := executeRequest(req)
-
-	resp1, err := formatResponse(response)
-	if err != nil {
-		return "", err
-	}
-
-	return resp1.Data.(string), nil
-}
-
-func formatResponse(r *httptest.ResponseRecorder) (handling.APIResponseJSON, error) {
-	response := handling.APIResponseJSON{}
-
-	body, err := ioutil.ReadAll(r.Body)
-	if err != nil {
-		return response, err
-	}
-
-	err = json.Unmarshal(body, &response)
-	return response, err
-}
-
-func formatData(i interface{}) []map[string]interface{} {
-	r := make([]map[string]interface{}, 0)
-	ilist := i.([]interface{})
-	for _, v := range ilist {
-		n := v.(map[string]interface{})
-		r = append(r, n)
-	}
-	return r
-}
-
-func insertUsers(d *models.Datasource, s []string) {
-	for _, v := range s {
-		u := models.Item{
-			"username":  v,
-			"password":  fmt.Sprintf("password_%s", v),
-			"email":     fmt.Sprintf("%s@example.com", v),
-			"firstname": v,
-			"lastname":  "User"}
-		d.InsertUser(u)
-	}
-}
-
-func clearCollections() {
-	sess := a.DB.Session
-	sess.DB("chmgt_test").C("Users").RemoveAll(nil)
-}
